refactor: project structure to use modules folder
