+++ src/modules/teams/index.ts
@@ -1,0 +1,2 @@
+export { TeamOverview } from './routes/TeamOverview';
+export { Teams } from './routes/Teams';
