+++ src/modules/teams/types/index.ts
@@ -1,0 +1,10 @@
+export interface Teams {
+  id: string;
+  name: string;
+}
+
+export interface TeamOverview {
+  id: string;
+  teamLeadId: string;
+  teamMemberIds: string[];
+}
