+++ src/modules/users/index.ts
@@ -1,0 +1,1 @@
+export { UserOverview } from './routes/UserOverview';
