+++ src/modules/users/routes/UserOverview.tsx
@@ -1,10 +1,10 @@
 import { useLocation } from 'react-router-dom';
 
-import { UserData } from '@/types';
+import Card from '@/components/Card';
+import { Container } from '@/components/GlobalComponents';
+import Header from '@/components/Header';
 
-import Card from '../components/Card';
-import { Container } from '../components/GlobalComponents';
-import Header from '../components/Header';
+import { UserData } from '../types';
 
 const mapU = (user: UserData) => {
   const columns = [
@@ -24,7 +24,7 @@ const mapU = (user: UserData) => {
   return <Card columns={columns} hasNavigation={false} navigationProps={user} />;
 };
 
-const UserOverview = () => {
+export const UserOverview = () => {
   const location = useLocation();
   return (
     <Container>
@@ -33,5 +33,3 @@ const UserOverview = () => {
     </Container>
   );
 };
-
-export default UserOverview;
