+++ src/modules/users/types/index.ts
@@ -1,0 +1,8 @@
+export interface UserData {
+  id: string;
+  firstName: string;
+  lastName: string;
+  displayName: string;
+  location: string;
+  avatar: string;
+}
